feat: enhance use case decorator to support async functions
